Split CLI option groups into helper modules
